+++ packages/custom_lint/CHANGELOG.md
@@ -1,3 +1,7 @@
+## Unreleased patch
+
+- Added support for `--fix`
+
 ## 0.5.11 - 2024-01-27
 
 - Added support for `analysis_options.yaml` that are nt at the root of the project (thanks to @mrgnhnt96)